MillasViajeras/requirements.txt
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
colorama==0.4.4
debugpy==1.5.1
decorator==5.1.1
dj-database-url==0.5.0
Django==4.0.6
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
entrypoints==0.4
executing==0.8.3
gunicorn==20.1.0
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
numpy==1.22.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.28
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pyzmq==22.3.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
tzdata==2022.1
wcwidth==0.2.5
whitenoise==6.2.0